+++ pkg/util/validation/user_limits_handler.go
@@ -17,7 +17,7 @@ type UserLimitsResponse struct {
 	// Write path limits
 	IngestionRate             float64 `json:"ingestion_rate"`
 	IngestionBurstSize        int     `json:"ingestion_burst_size"`
-	IngestionBurstFactor      int     `json:"ingestion_burst_factor"`
+	IngestionBurstFactor      float64 `json:"ingestion_burst_factor"`
 	MaxGlobalSeriesPerUser    int     `json:"max_global_series_per_user"`
 	MaxGlobalSeriesPerMetric  int     `json:"max_global_series_per_metric"`
 	MaxGlobalExemplarsPerUser int     `json:"max_global_exemplars_per_user"`
